Enable optimizations fix warnings and one buffer problem
Somehow it appeared that there were not -O2 or alike settings via CFLAGS.
Without those my experience is to get slower code than desired. I suggest
to enable those.

I furthermore enabled -Wall and -W to enable the compiler help to figure
out potential problems.

I fixed the resulting warnings, e.g. unused variables or parameters.
gcc specific, I hope the attribute stuff is no problem.

And last but not least there is really one problem with a command buffer
which is used with sprintf, but was just declared as char * with no
memory behind it as far as I could see.